refactor: change directory, add new package for StatAnalyzeDecoder
