+++ src/main/java/ru/javarush/cryptoanalyzer/konovalov/Runner.java
@@ -5,8 +5,8 @@ String[] args[] for this program
 String [0] - number of operationMode
 Caesar Encoder - {0, file.txt, resultFile.txt, key from 1 to 50}
 Caesar Decoder - {1, encryptedFile.txt, resultFile.txt, key from 1 to 65 = getCryptAlphabetArrayLength()-1}
-BruteForceDecoder - {2, encryptedFile.txt, resultFile.txt, example.txt}
-Statistical Analysis Decoder {3, encryptedFile.txt, resultFile.txt, example.txt}
+BruteForceDecoder - {2, encryptedFile.txt, resultFile.txt, Vocabulary.txt}
+Statistical Analysis Decoder {3, encryptedFile.txt, resultFile.txt,example.txt }
 Exit {4}
  */
