--- src/main/java/ru/javarush/cryptoanalyzer/konovalov/strategy/StatAnalyzeDecoder.java
@@ -1,9 +1,0 @@
-package ru.javarush.cryptoanalyzer.konovalov.strategy;
-
-public class StatAnalyzeDecoder implements Actionable{
-
-    public void action(String[] args){
-
-    }
-
-}
